+++ json/assembly/index.ts
@@ -1,18 +1,14 @@
-// import { JSONSerializer } from './serializer';
-// import { JSONDeserializer } from './deserialize';
 import { Serial } from '@serial-as/core';
-import { ValueDeserializer, ValueSerializer } from './value'
+import { ValueSerializer } from './serializer'
+import { ValueDeserializer } from './deserializer'
+
 import { JSON } from "assemblyscript-json";
 
-// TODO figure out complier issue that prevented these exports.
-// export {JSONSerializer, JSONDeserializer}
 
-
-  /**
-   * 
-   * @param s string or JSON.Value
-   * @returns 
-   */
+/*
+ * @param s string or JSON.Value
+ * @returns 
+ */
 export function parse<T, From = string>(s: From): T {
     return ValueDeserializer.decode<T, From>(s);
   }
