--- json/assembly/value/index.ts
@@ -1,3 +1,0 @@
-export * from "./serializer"
-export * from "./deserializer"
-
