+++ package.json
@@ -15,7 +15,7 @@
   "scripts": {
     "test": "asp",
     "test:borsh": "asp -f bencode",
-    "test:json": "asp -f jencode",
+    "test:json": "asp -f jvalue",
     "build": "yarn workspace @serial-as/transform build",
     "publish": "lerna publish"
   },
